moved ObjectiveAndGoals to store and renamed buildGoal
